Move agentpoolonly API model to api package
