+++ pkg/api/agentPoolOnlyApi/v20170831/const.go
