+++ pkg/api/agentPoolOnlyApi/v20170831/types_test.go
