+++ pkg/api/convertertoapi.go
@@ -1,7 +1,7 @@
 package api
 
 import (
-	"github.com/Azure/acs-engine/pkg/agentPoolOnlyApi/v20170831"
+	"github.com/Azure/acs-engine/pkg/api/agentPoolOnlyApi/v20170831"
 	"github.com/Azure/acs-engine/pkg/api/v20160330"
 	"github.com/Azure/acs-engine/pkg/api/v20160930"
 	"github.com/Azure/acs-engine/pkg/api/v20170131"
